requirements.txt
alembic==1.4.3
aniso8601==6.0.0
Babel==2.9.0
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
ecdsa==0.14.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.0.0
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.2.0
pytz==2019.1
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.4
Werkzeug==1.0.1
